requirements.txt
argh==0.26.2
Django==2.1.2
django-appconf==1.0.2
django-compressor==2.2
django-hosts==3.0
djangorestframework==3.9.0
Markdown==3.0.1
pathtools==0.1.2
pendulum==2.0.4
psycopg2-binary==2.7.5
Pygments==2.2.0
pygments-style-github==0.4
pymdown-extensions==6.0
python-dateutil==2.7.5
python-frontmatter==0.4.2
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
rcssmin==1.0.6
rjsmin==1.0.12
ruamel.yaml==0.15.75
six==1.11.0
tqdm==4.28.1
watchdog==0.9.0
